Corrections to mistakes and an addition to vimrcs

+++ newmachinesetup.sh
@@ -6,19 +6,23 @@
 #############
 # BASH SETUP
 #############
-cp ~/.bashrc ~/.bashrc.orig
+SETUP_DIR=$(pwd);
+
+if [ -f ~/.bashrc ]; then
+	cp ~/.bashrc ~/.bashrc.orig
+fi
 if [ -f ~/.bash_aliases ]; then
     cp ~/.bash_aliases ~/.bash_aliases.orig
 fi
-cp bashrc ~/.bashrc
-cp bash_aliases ~/.bash_aliases
+cp $SETUP_DIR/dotfiles/bashrc ~/.bashrc
+cp $SETUP_DIR/dotfiles/bash_aliases ~/.bash_aliases
 
 
 ################
 # ESSESSENTIALS
 ################
 sudo apt-get update
-sudo apt-get install vim git python3
+sudo apt install vim git python3
 if [ ! -d ~/Downloads ]; then
     mkdir ~/Downloads
 fi
@@ -43,17 +47,16 @@ read FULLNAME
 git config --global user.name $FULLNAME
 echo 'Please enter email for git:'
 read GITEMAIL
-git config --global user.name $GITEMAIL
+git config --global user.email $GITEMAIL
 git config --global core.editor vim
 
 #############
 # Python
 #############
-cd ~/Downloads
 wget -O ~/Downloads/get-pip.py https://bootstrap.pypa.io/get-pip.py
-sudo python3 get-pip.py
+sudo python3 ~/Downloads/get-pip.py
 sudo pip install virtualenv virtualenvwrapper
-sudo pip install --user numpy
+sudo pip install --user numpy pandas matplotlib
 pip completion -b >> ~/.profile
 
 ################
@@ -61,26 +64,27 @@ pip completion -b >> ~/.profile
 ################
 git clone --depth=1 git://github.com/amix/vimrc.git ~/.vim_runtime
 sh ~/.vim_runtime/install_awesome_vimrc.sh
+cat $SETUP_DIR/dotfiles/vimrcadditions >> ~/.vim_runtime/vimrcs/basic.vim
 
 ###############
 # Sublime Text
 ###############
 wget -qO - https://download.sublimetext.com/sublimehq-pub.gpg | sudo apt-key add -
-sudo apt-get install apt-transport-https
+sudo apt install apt-transport-https
 echo "deb https://download.sublimetext.com/ apt/stable/" | sudo tee /etc/apt/sources.list.d/sublime-text.list
-sudo apt-get update
-sudo apt-get install sublime-text
+sudo apt update
+sudo apt install sublime-text
 
 ##########
 # Tmux
 ##########
 sudo apt install tmux
-cp tmux.conf ~/.tmux.conf
+cp $SETUP_DIR/dotfiles/tmux.conf ~/.tmux.conf
 
 #################
 # USEFULL THINGS
 #################
-sudo apt-get install evince
+sudo apt install evince tree
 
 ##################
 # UBUNTU ARCTHEME
@@ -88,18 +92,22 @@ sudo apt-get install evince
 if [[ `uname -v` == *"Ubuntu"* ]]; then
     #Theme
     sudo apt-get install autoconf automake pkg-config libgtk-3-dev gtk2-engine-murrine gnome-tweak-tool
+    cd ~/Downloads
     git clone https://github.com/horst3180/arc-theme --depth 1
     cd arc-theme
     ./autogen.sh --prefix=/usr --disable-light --disable-darker
     sudo make install
 
     #Icons
+    cd ~/Downloads
     git clone https://github.com/horst3180/arc-icon-theme --depth 1 && cd arc-icon-theme
     ./autogen.sh --prefix=/usr
     sudo make install
     sudo add-apt-repository ppa:moka/daily
     sudo apt-get install moka-icon-theme
     gnome-tweak-tool
+
+    cd $START_DIR
 fi
 
 ##############################
